+++ tilemap.h
@@ -15,7 +15,7 @@
 #define MKTILE(set, id) (((set) << 10) | (id))
 #define NOTILE UINT16_MAX
 
-typedef uint16_t tile_t;
+typedef uint32_t tile_t;
 
 struct tileset {
     struct image img;
@@ -23,10 +23,10 @@ struct tileset {
     size_t refc;
     struct tile {
         struct rect pos;
-        int16_t origin_x;
-        int16_t origin_y;
+        int32_t origin_x;
+        int32_t origin_y;
         /* Next animation frame */
-        int32_t next_frame;
+        uint32_t next_frame;
     } *tiles;
 };
 
@@ -36,8 +36,8 @@ struct tilemap {
     struct tileset **sets;
     size_t width;
     size_t height;
-    int16_t tile_width;
-    int16_t tile_height;
+    int32_t tile_width;
+    int32_t tile_height;
     uint32_t *dirty;
     double scale;
     tile_t tiles[];
@@ -47,15 +47,15 @@ struct tilemap {
 struct tileset *create_tileset(const char *path, struct tile *tiles, size_t ntiles);
 void unref_tileset(struct tileset *);
 void ref_tileset(struct tileset *);
-void tileset_queue_tile(struct image dst, struct tileset *set, tile_t tile, int16_t x, int16_t y, double scale);
+void tileset_queue_tile(struct image dst, struct tileset *set, tile_t tile, int32_t x, int16_t y, double scale);
 
-struct tilemap *create_tilemap(size_t width, size_t height, int16_t tile_width, int16_t tile_height, struct tileset **sets, size_t nsets) ;
+struct tilemap *create_tilemap(size_t width, size_t height, int32_t tile_width, int16_t tile_height, struct tileset **sets, size_t nsets) ;
 void free_tilemap(struct tilemap *map);
 tile_t tilemap_add_tileset(struct tilemap *map, struct tileset *tileset);
-void tilemap_queue_draw(struct image dst, struct tilemap *map, int16_t x, int16_t y);
-tile_t tilemap_set_tile(struct tilemap *map, int16_t x, int16_t y, int16_t layer, tile_t tile);
+void tilemap_queue_draw(struct image dst, struct tilemap *map, int32_t x, int16_t y);
+tile_t tilemap_set_tile(struct tilemap *map, int32_t x, int16_t y, int16_t layer, tile_t tile);
 void tilemap_set_scale(struct tilemap *map, double scale);
-tile_t tilemap_get_tile(struct tilemap *map, int16_t x, int16_t y, int16_t layer);
+tile_t tilemap_get_tile(struct tilemap *map, int32_t x, int16_t y, int16_t layer);
 void tilemap_animation_tick(struct tilemap *map);
 void tilemap_random_tick(struct tilemap *map);
 void tilemap_refresh(struct tilemap *map);
