+++ util.h
@@ -34,32 +34,32 @@
 
 typedef uint32_t color_t;
 struct rect {
-    int16_t x;
-    int16_t y;
-    int16_t width;
-    int16_t height;
+    int32_t x;
+    int32_t y;
+    int32_t width;
+    int32_t height;
 };
 
-inline static struct rect rect_scale_up(struct rect rect, int16_t x_factor, int16_t y_factor) {
+inline static struct rect rect_scale_up(struct rect rect, int32_t x_factor, int16_t y_factor) {
     rect.x *= x_factor;
     rect.y *= y_factor;
     rect.width *= x_factor;
     rect.height *= y_factor;
     return rect;
 }
-inline static struct rect rect_scale_down(struct rect rect, int16_t x_factor, int16_t y_factor) {
+inline static struct rect rect_scale_down(struct rect rect, int32_t x_factor, int16_t y_factor) {
     rect.x /= x_factor;
     rect.y /= y_factor;
     rect.width /= x_factor;
     rect.height /= y_factor;
     return rect;
 }
-inline static struct rect rect_shift(struct rect rect, int16_t x_off, int16_t y_off) {
+inline static struct rect rect_shift(struct rect rect, int32_t x_off, int16_t y_off) {
     rect.x += x_off;
     rect.y += y_off;
     return rect;
 }
-inline static struct rect rect_resize(struct rect rect, int16_t x_off, int16_t y_off) {
+inline static struct rect rect_resize(struct rect rect, int32_t x_off, int16_t y_off) {
     rect.width += x_off;
     rect.height += y_off;
     return rect;
